handle timeouts properly
We use alarms and poll timeouts to properly deal with the case of a
server gone AWOL. Should prevent needing to kill the damn thing every
time the primary is rebooted. use 10 seconds for timeout, however, as
the 2 seconds specified in original uSynergy is excessively short and
prone to errors.