a proper attempt at this -- it *almost* works
monitoring works

Now all that is left is wl-copy removal

almost works -- dies for some reason

fully native clipboard support -- no more wl-clip

update README

...this kind of sucks now

fork to send, ignore SIGCHLD

We weren't doing anything with SIGCHLD anyway, and the forking approach
means we don't ever run into the situation of hanging on a clipboard
loop, which unfortunately can't really be terminated with the existing
set of protocols (i.e. we cannot know who sent the offer, so we must
receive. And we aren't doing our job if we ignore offers. So...)

shorter clipboard updater timeout

well this is hopeless now isnt it