fix protocol version of wl_output
wl_output now offers names and descriptions, but we use xdg_output for
that. Fix the protocol version at 2 for the time being (should solve #28)
pending further evaluation of whether xdg_output is needed at all

And bump to 0.0.6, because I don't like the idea a release not working
with the latest sway.