clarify uintput error
The uinput file descriptors are opened immediately at startup in order
to reduce the amount of time spent with elevated privileges, which are
required if someone is using the uinput backend. Add a note that it is
normal for this to fail if someone is using another backend, as it will
be, and will pose no problem. In response to #28