support mixing maps of synergy and raw keys
This should aid in fixing #35, as Windows seems to be sending identical
buttons for different keys, but distinct synergy identifiers.

Also support setting a raw keycode for idle... which goes here becuase
it will solve the problem of HYPR on uinput triggering the play key.

move debug log to include wlRawKey

update readme