allow disabling idle inhibition entirely
Synergy upstream no longer sends the screensaver sync messages, leading
to issues with screenlockers as in
https://github.com/r-c-f/waynergy/issues/49
so we add an `idle-inhibit/enable' bool option to enable or disable
inhibition. Barrier/input-leap still seem to support this feature and I
use it, so I'll default to leaving it on for now.