handle fatal errors more gracefully
Previously, calling Exit or ExitOrRestart to deal with a fatal Wayland
error would attempt to call the cleanup() function... which would then
try to close the Wayland connection gracefully, which would then trigger
another fatal error and crash waynergy in the process. By defining a new
status enum we can indicate when an error originates as a fatal wayland
error and avoid this. Fixes our half of #71.